+++ src/Core/Timer.cpp
@@ -17,10 +17,10 @@ void core::Timer::restart()
 	m_start_time = std::chrono::steady_clock::now();
 }
 
-float core::Timer::getElapsedSeconds()
+double core::Timer::getElapsedSeconds()
 {
 	std::chrono::steady_clock::time_point end_time = std::chrono::steady_clock::now();
-	return std::chrono::duration<float>(end_time - m_start_time).count();
+	return std::chrono::duration<double>(end_time - m_start_time).count();
 }
 
 } //namespace eng
